+++ config_template.py
@@ -60,7 +60,7 @@ PASSLIB_CONTEXT = {
 
 # WebHCat settings
 WEBHCAT_BASE_URL = 'http://localhost:50111/templeton/v1/'
-WEBHCAT_CALLBACK_URL = 'http://localhost/rest/queries/callback/{id}'
+WEBHCAT_CALLBACK_NETLOC = 'localhost'
 WEBHCAT_SCRIPT_COMMON = textwrap.dedent("""\
     SET hive.exec.copyfile.maxsize=1073741824;
     """
@@ -85,11 +85,12 @@ MAIL_PASSWORD = None
 MAIL_DEFAULT_SENDER = 'CosmoHub <cosmohub@pic.es>'
 
 MAIL_SUBJECTS = {
-    'acls_change' : u'Group membership updated',
+    'acls_updated' : u'Group membership updated',
+    'acls_request' : u'Group membership request',
     'email_confirm' : u'Confirm your email address',
     'password_reset' : u'Reset your CosmoHub password',
     'query_ready' : u'Your catalog (#{id}) is ready!',
-    'user_register' : u'Hi {name}, welcome to CosmoHub!',
+    'user_registered' : u'Hi {name}, welcome to CosmoHub!',
 }
 
 RECAPTCHA_ENABLED = True