+++ cosmohub/api/rest/user.py
@@ -3,40 +3,62 @@ import urllib
 import urlparse
 import werkzeug.exceptions as http_exc
 
-from flask import g, current_app, render_template, request
-from flask_restful import Resource, reqparse, marshal
+from flask import (
+    current_app,
+    g,
+    render_template,
+    request,
+)
+from flask_restful import (
+    marshal,
+    reqparse,
+    Resource,
+)
 from sqlalchemy.orm import (
     joinedload,
     undefer_group,
 )
 from sqlalchemy.sql import func
 
-from cosmohub.api import db, api_rest, mail, recaptcha
+from cosmohub.api import (
+    api_rest,
+    db,
+    mail,
+    recaptcha,
+)
 
 from .. import fields
 from ..database import model
-from ..database.session import transactional_session, retry_on_serializable_error
-from ..security import adler32, auth_required, Privilege, Token
+from ..database.session import (
+    retry_on_serializable_error,
+    transactional_session,
+)
+from ..security import (
+    adler32,
+    auth_required,
+    Privilege,
+    Token,
+)
 
 log = logging.getLogger(__name__)
 
 class UserItem(Resource):
-    @auth_required(Privilege(['user']))
+    @auth_required(Privilege('/user'))
     def get(self):
         with transactional_session(db.session, read_only=True) as session:
-            user = session.query(
-                model.User
-            ).options(
-                joinedload('acls')
-            ).filter_by(
-                id=g.session['user'].id
-            ).one()
-            
             groups = session.query(
                 model.Group
             ).options(
                 undefer_group('text'),
             ).all()
+            
+            user = session.query(
+                model.User
+            ).options(
+                joinedload('acls'),
+            ).filter_by(
+                id=g.session['user'].id
+            ).one()
             
             user.groups = []
             for group in groups:
@@ -54,7 +76,7 @@ class UserItem(Resource):
             
             return marshal(user, fields.User)
     
-    @auth_required( Privilege(['user'], ['fresh']) | Privilege(['password_reset']) )
+    @auth_required( Privilege('/user') | Privilege('/password_reset') )
     def patch(self):
         parser = reqparse.RequestParser()
         parser.add_argument('name', store_missing=False)
@@ -65,12 +87,14 @@ class UserItem(Resource):
         attrs = parser.parse_args(strict=True)
 
         with transactional_session(db.session) as session:
-            user = session.query(model.User).filter_by(
+            user = session.query(
+                model.User
+            ).filter_by(
                 id=g.session['user'].id
             ).with_for_update().one()
             
-            privilege = Privilege(['user'], ['fresh'])
-            privilege |= Privilege(['password_reset'], [adler32(user.password.hash)])
+            privilege = Privilege('/user')
+            privilege |= Privilege('/password_reset/{0}'.format(adler32(user.password.hash)))
             if not privilege.can(g.session['privilege']):
                 raise http_exc.Unauthorized
             
@@ -91,7 +115,7 @@ class UserItem(Resource):
                 
                 token = Token(
                     user, 
-                    Privilege(['email_confirm'], [adler32(user.email)]), 
+                    Privilege('/email_confirm/{0}'.format(adler32(user.email))), 
                     expires_in=current_app.config['TOKEN_EXPIRES_IN']['email_confirm'],
                 )
                 url = urlparse.urljoin(request.environ['HTTP_REFERER'], attrs['redirect_to'])
@@ -109,9 +133,10 @@ class UserItem(Resource):
         parser.add_argument('name', required=True)
         parser.add_argument('email', required=True)
         parser.add_argument('password', required=True)
-        parser.add_argument('requested_groups', required=True, action='append')
+        parser.add_argument('requested_groups', default=[], action='append')
         parser.add_argument('recaptcha', required=True)
-        parser.add_argument('redirect_to', required=True)
+        parser.add_argument('email_confirm_route', required=True)
+        parser.add_argument('acl_update_route', required=True)
 
         attrs = parser.parse_args(strict=True)
 
@@ -119,14 +144,21 @@ class UserItem(Resource):
             raise http_exc.Unauthorized('Captcha invalid')
         del attrs['recaptcha']
         
-        url = urlparse.urljoin(request.environ['HTTP_REFERER'], attrs['redirect_to'])
-        del attrs['redirect_to']
+        email_confirm_url = urlparse.urljoin(request.environ['HTTP_REFERER'], attrs['email_confirm_route'])
+        del attrs['email_confirm_route']
+        
+        acl_update_url = urlparse.urljoin(request.environ['HTTP_REFERER'], attrs['acl_update_route'])
+        del attrs['acl_update_route']
         
         with transactional_session(db.session) as session:
             requested_groups = attrs.pop('requested_groups')
             
-            groups = session.query(model.Group).filter(
+            groups = session.query(
+                model.Group
+            ).filter(
                 model.Group.name.in_(requested_groups),
+            ).options(
+                joinedload('users_admins'),
             ).with_for_update().all()
 
             if len(groups) != len(requested_groups):
@@ -141,17 +173,47 @@ class UserItem(Resource):
             
             token = Token(
                 user, 
-                Privilege(['email_confirm'], [adler32(user.email)]), 
+                Privilege('/email_confirm/{0}'.format(adler32(user.email))), 
                 expires_in=current_app.config['TOKEN_EXPIRES_IN']['email_confirm'],
             )
             
-            url += '?' + urllib.urlencode({ 'auth_token' : token.dump() })
+            email_confirm_url += '?' + urllib.urlencode({ 'auth_token' : token.dump() })
             
             mail.send_message(
-                subject = current_app.config['MAIL_SUBJECTS']['user_register'],
+                subject = current_app.config['MAIL_SUBJECTS']['user_registered'],
                 recipients = [user.email],
-                body = render_template('mail/user_register.txt', user=user, url=url),
-                html = render_template('mail/user_register.html', user=user, url=url),
+                body = render_template(
+                    'mail/user_registered.txt',
+                    user=user,
+                    url=email_confirm_url
+                ),
+                html = render_template(
+                    'mail/user_registered.html',
+                    user=user,
+                    url=email_confirm_url
+                ),
+            )
+            
+            recipients = set()
+            for group in groups:
+                for user in group.users_admins:
+                    recipients.add(user.email)
+            
+            mail.send_message(
+                subject = current_app.config['MAIL_SUBJECTS']['acl_request'],
+                recipients = recipients,
+                body = render_template(
+                    'mail/acl_request.txt',
+                    user=user,
+                    groups=groups,
+                    url=acl_update_url,
+                ),
+                html = render_template(
+                    'mail/acl_request.html',
+                    user=user,
+                    groups=groups,
+                    url=acl_update_url,
+                ),
             )
             
             g.session['track']({
@@ -160,10 +222,10 @@ class UserItem(Resource):
                 'ea' : 'register',
                 'el' : user.id,
             })
-        
+            
             return marshal(user, fields.User), 201
 
-    @auth_required(Privilege(['user'], ['fresh']))
+    @auth_required(Privilege('/user'))
     def delete(self):
         @retry_on_serializable_error
         def delete_user(user_id):
@@ -188,14 +250,14 @@ class UserItem(Resource):
 api_rest.add_resource(UserItem, '/user')
 
 class UserEmailConfirm(Resource):
-    @auth_required(Privilege(['email_confirm']))
+    @auth_required(Privilege('/email_confirm'))
     def get(self):
         with transactional_session(db.session) as session:
             user = session.query(model.User).filter_by(
                 id=g.session['user'].id
             ).with_for_update().one()
             
-            privilege = Privilege(['email_confirm'], [adler32(user.email)])
+            privilege = Privilege('/email_confirm/{0}'.format(adler32(user.email)))
             if not privilege.can(g.session['privilege']):
                 raise http_exc.Forbidden
             
@@ -235,7 +297,7 @@ class UserPasswordReset(Resource):
             
             token = Token(
                 user, 
-                Privilege(['password_reset'], [adler32(user.password.hash)]), 
+                Privilege('/password_reset/{0}'.format(adler32(user.password.hash))), 
                 expires_in=current_app.config['TOKEN_EXPIRES_IN']['password_reset'],
             )